extend_volume of libvirt/volume/iscsi should not use device_path
The connection_info['data']['device_path'] field is not always
available. In cases that it was not available, it would cause
the debug code to raise a KeyError instead of proceeding.

Other similar debug messages in the same file do not include
device_path. As a simple fix, just drop the device_path from
the log.

Closes-Bug: #1936439

Change-Id: Id0539d2ee909d86ffef07ae566697db8ae0f83b4
Signed-off-by: Mark Mielke <mark.mielke@gmail.com>