doc: allow_resize_to_same_host needs nova-api restarted
[DEFAULT]allow_resize_to_same host is a bit weird. It controls
scheduling, is set in the [DEFAULT] section, but needs nova-api to be
restarted for changes to take effect (the API layer calls
compute_api.resize(), and the allow_resize_to_same_host check is
performed there, and not in the scheduler). This patch adds a small
blurn to the config help text to explain this.

Change-Id: I5393dfaeaf123d28bb1889635792ee1ece2c5295