Allow deletion of compute service with no compute nodes
Consider the following situation:

- Using the Ironic virt driver
- Replacing (so removing and re-adding) all baremetal nodes
  associated with a single nova-compute service

The update resources periodic will have destroyed the compute node
records because they're no longer being reported by the virt driver.
If we then attempt to manually delete the compute service record, the
datbase layer will raise an exception, as there are no longer any
compute node records for the host. Previously, this exception would
get bubbled up as an error 500 in the API. This patch catches it and
allows service deletion to complete succefully.

Closes bug: 1860312
Change-Id: I2f9ad3df25306e070c8c3538bfed1212d6d8682f