compute: Query the service group API within check_instance_host
The check_instance_host decorator when called with the check_is_up kwarg
is supposed to ensure that the compute service associated with a given
instance is fully up. Previously it used the HostStatus returned by
get_instance_host_status to determine this allowing either a state of UP
or MAINTENANCE.

However get_instance_host_status currently returns a HostStatus of
MAINTENANCE if the service is simply disabled regardless of the
underlying service being active or not. As a result of this and as set
out in bug #1938326 the check_instance_host decorator will pass allowing
calls to resize an instance to continue and later fail when RPC requests
to the instance host eventually fail.

To resolve this check_instance_host is now updated to only use the
service group API to determine if the underlying service is up.

Future work will look at updating the meaning of the MAINTENANCE
HostStatus to mean disabled but up. As this will potentially
change values returned in the API to admins doing an instance show this
will need to be done behind a new microversion to ensure backward
compatibility with the previous behaviour and meaning of the HostStatus.

Closes-Bug: #1938326
Change-Id: I45fd908b458e00b5e18848f624043ceb88b04825