Reproducer unit test for bug 1860312
Consider the following situation:

- Using the Ironic virt driver
- Replacing (so removing and re-adding) all baremetal nodes
  associated with a single nova-compute service

The update resources periodic will have destroyed the compute node
records because they're no longer being reported by the virt driver.
If we then attempt to manually delete the compute service record, the
datbase layer will raise an exception, as there are no longer any
compute node records for the host. This exception gets bubbled up as
an error 500 in the API. This patch adds a unit test to demonstrate
this.

Related bug: 1860312
Change-Id: I03eec634b25582ec9643cacf3e5868c101176983