Provide the mdev class for every PCI device
We would want to modify the resource class if the mdev_class option
is used.
So, we provide two methods for getting the supported resource
classes, one for knowing which RC is used for a specific PCI ID
and the other for getting all the related RCs.

Partially-Implements: blueprint generic-mdevs
Change-Id: I0268ba0d70002b2abce553f821e7b6d55b260842