Provide and use other RCs for mdevs if needed
Now that we have helper methods for knowing which RC we should use
for a specific mdev-supported device, use those RCs for both the
related inventories and for looking at the right allocations.

NOTE(sbauza): Given we no longer provide VGPU resources to a root
Resource Provider since Ussuri, we don't need to support reshapes
for allocations that are "generic mdevs" in the root RP, hence the
reshape methods not being modified.

Change-Id: I0bb93ab0952204e06a50829c4a66b03bddb5be45
Partially-Implemented: blueprint generic-mdevs