func: Increase rpc_response_timeout in TestMultiCellMigrate tests
This was previously set really low to 1 second that was leading to more
involved flows such as test_delete_while_in_verify_resize_status timing
out when the target calls the conductor to confirm the resize on the
source.

This change simply increases the timeout in the test but we might want
to think about moving this call over to rpc_long_timeout this could be
an issue in real world deployments.

Closes-Bug: #1938021
Change-Id: Ibba2d1506a0b026d35d7bf35384ec6439f438b01