Add a regression test for bug 1939545
Assert the behaviour of the libvirt driver during pre_live_migration
with instances that have block based volumes attached.

Bug #1939545 covering the case where the returned path from os-brick
isn't being saved into the connection_info of the associated bdm in
Nova.

The cinder fixture is extended as part of this change to offer
connection_info listing the iscsi volume type in order to correctly
assert the behaviour of the associated libvirt volume driver.

Related-Bug: #1939545
Change-Id: I3f1073d70e4332737e33f3c9c00cf129d0ef76b4