fake: Ensure need_legacy_block_device_info returns False
The legacy block_device_info format has not been used for a while in the
virt drivers and the Fake virt drivers shouldn't be using it anymore.

Some compute tests have been updated as they were previously asserting
the legacy format incorrectly. TODOs are also left to remove support for
the legacy format in a follow up change.

Change-Id: I066fb5617bae944a03273049117fc2b5bf329283