Vmware: Refactor server groups
Previously vmops was calling the private function
`vm_util._get_server_groups` and that function would access to nova.object.
All other code seem to work more the way, that the nova.object retrieval is
responsibility of a VmOps (or VolumeOps) instance. `vm_util` only works
with vmware-objects (plus passed instances).

Additionally, `update_cluster_placement` was called in
`VMwareVMOps.build_virtual_machine`, which is also called for creating
a virtual machine for a vm-template out of an image.
In that case, the server-groups of the instance requiring the image
would be wrongly applied to the vm-template of the image.

(cherry picked from commit 475fdb14fde79aea7c8098741b1b9c0c00fecc7b)