db: Handle parameters in DB strings
As part of 'nova.db.migration', it is necessary to override the
'sqlalchemy.url' config option used by alembic. We were doing this but
we weren't handling the fact that we could receive encoded URL strings.
This causes issues for the ConfigParser instead used under the hood, as
noted in the alembic docs [1]:

  value – the value. Note that this value is passed to ConfigParser.set,
  which supports variable interpolation using pyformat (e.g.
  %(some_value)s). A raw percent sign not part of an interpolation
  symbol must therefore be escaped, e.g. %%. The given value may refer
  to another value already in the file using the interpolation format.

Resolve the issue by escaping % with %%. The engine.url is always
encoded therefore this can be done unconditionally.

[1] https://alembic.sqlalchemy.org/en/latest/api/config.html#alembic.config.Config.set_main_option

Closes-Bug: #1940555
Change-Id: I74de55107a80af13df348f2bce49415b08028746
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Co-Authored-By: Sean Mooney <work@seanmooney.info>