api: Log actual number of UUIDs queried
When retrieving a server group, we log in debug log how many instances
we query each cell for. This used the wrong variable, always showing the
same number for all cells - the number of members in the server group -
instead of showing the number of UUIDs we query the cell for.

Change-Id: Ib2308653e1894e1ab9462eec01b19f1a038552f0