Reproduce bug 1941005
We don't update instance.request_spec when we resize to a new flavor.
If the flavor's extra specs change, this is not reflected in
instance.request_spec.

This generally goes unnoticed when *adding* stuff to an instance. The
scheduler will still find a host based on the old request spec. The
resource claim on the destination uses the flavor and image data
directly (not the request spec) so the claim will correctly fail if
the destination does not have the new things that the instance wants.

When *removing* stuff form an instance this becomes problematic. The
scheduler will attempt to schedule based on the old request spec,
which could end up with NoValidHost if no host exists that has the
*old* stuff the instance wanted, despite the new flavor not wanting
any of that stuff.

Related bug: 1941005

Change-Id: I8c052a6526819834d79d2b5027050f1bdf36638f