Remove module level caching
This change removes the module level caches
_MIGRATE_REPO and _ALEMBIC_CONF from nova.db.migration

These caches are not correctly reset in current testing
and dont currently provide much value as it is atypical
for db_sync or db_version to be invoked multiple times outside
of tests within the same process. As such the caches will typically
be empty.

Change-Id: I7a59c29f27dc16a0a2b3bb688a52da3c962a0b1b