Parse alias from domain hostdev
In I16e7df6932bb7dff243706ee49338ba6b3782085 we missed that
LibvirtConfigGuestHostdevPCI is not a child class of
LibvirtConfigGuestInterface and therefore we missed parsing out the
alias field from the domain xml for hostdevs. The new libvirt driver
device detach logic[1] uses the alias as the identifier towards libvirt
so now hostdevs cannot be detached. This patch parses out the alias
field to fix the issue.

Closes-Bug: #1942345
Related-Bug: #1882521

[1] https://review.opendev.org/q/topic:bug/1882521

Change-Id: I30d30a772475cb82d0fd088f14a54a35646bd1dc