Avoid excessive sleep in vmware unit test
The patch Id6e2be2b407eb08e269c35e0f89d235c67a1ab99 added a retry
decorator on create_vm in vmware driver. Such decorator makes the
negative unit test do excessive sleeps and retries. The test eventually
stopped by the test executor with a Timeout exception but the test was
modified to assert that as a success. This makes the test pass after
160 seconds.

This patch mocks the sleep behind the retry decorator to make the test
fast and reverts the assert in the test.

Change-Id: I7c7e698bf559b69a2f09402c265e6e387c3b077a
Related-Bug: #1908408