Add more retries to TestMigrateFromDownHost tests
We still having occasional failures in TestMigrateFromDownHost tests
where the compute is not observed down after stopping the service. There
was a fix I61fded6f513c18dbd0248ed54fc58a318dc05416 that increased the
CONF.service_down_time to 6 seconds. However the
_wait_for_service_parameter() helper by default uses 10 retries with 0.5
sec delays. This can be pretty close to the 6 seconds of service
timeout. This patch increase the number of retires to 20 avoid unstable
test runs.

Change-Id: Iad970fa3aa10aaa2c831a064c085c06fdc88305e
Closes-Bug: #1940741