Fix nova-manage db version
When nova switched to alembic implementation was added to nova-manage db
version CLI to query the current db revision from alembic. However
multiple mistake was made.

The code called alembic_api.current[1] with an Engine object while that
call expects a Config object instead. This leads to but/1943436.

Also the same code expected that this call returns the revision. But
that call just prints the revision to the standard output instead.

So the implementations has been change from calling the alembic command
API which is mostly created for CLI consumption to
MigrationContext.get_current_revision() call that is intended to be used
as a python API instead.

[1] https://alembic.sqlalchemy.org/en/latest/api/commands.html#alembic.command.current

Co-Authored-By: Sean Mooney <smooney@redhat.com>
Closes-Bug: #1943436

Change-Id: I9fa7c03310c5bdb82e9a9c39727edb12eeae77f0