Add a way to trigger a server-group sync on the driver
We're going to change the API to allow updates of server-group members
and need to trigger a sync in the backend, whenever such a change
occurs. Therefore, we need a bunch of methods going through the whole
stack to call the driver.

We use a cast and not a call here, because we cannot let the API wait
for all of this to happen. If the cast gets lost somehow, a sync-loop
implemented in the driver will pick the change up, eventually

The API will have to supply a list of hosts to call, so we only sync the
group on the necessary hosts.

Change-Id: I00e012ed52ba9fd36b094ecf2dc86b023f2f5a21
(cherry picked from commit f3d2ad7e457ab0e47e63c82f09f91572ba04711f)