[stable-only] Set lower-constraints job as non-voting
This patch sets the failing lower-constraints job to non-voting until
final fix arrives for the failure.

With the latest setuptools (bundled into python virtualenv) the lower
constraints job started to fail on ussuri when it tries to install old
version of 'decorator'. The error message is:

  error in decorator setup command: use_2to3 is invalid.

The needed minimal version of 'decorator' would be 4.0.0, but it is
only needed if the latest version of setuptools (virtualenv) is used.
On old branches it would be good to pin setuptools (virtualenv),
otherwise we can face similar issues. Until the final decision / fix,
let's just set the lower-constraints job non-voting to unblock the
gate.

Change-Id: I01700882ca63f26975d4ecfd7d0a90ea95fe6614