Add regression test for bug #1943431
This regression test asserts the current broken behaviour of Nova during
a Cinder orchestrated volume migration that leaves the stashed
connection_info of the attachment pointing at a now deleted temporary
volume UUID used during the migration.

This is slightly different to the Nova orchestrated pure swap_volume
flow so an additional test is included to assert the current correct
behaviour there and to hopefully illustrate the differences better to
reviewers.

Numerous TODOs are also left in place for clean ups in the regression
test, Cinder fixture and integrated helpers.

Change-Id: Ic060523c30fd4dae086afcf7566452be0b2ff266