vmware: Do not sync "soft-affinity" server-groups
The replaced codde in cluster_util already ignored server-groups having
a policy of "soft-affinity", so we have to do it in the new
sync_server_group(), too.

The main reason is, that we wanted to give customers the possibility to
end up on the same cluster, without the need to end up on the same host,
too. We also cannot implement the "soft" part currently and  spawning a
VM will fail if the host is too full, as there are no non-mandatory
VM-to-VM rules.

Change-Id: I745ac6616eefd193ce8c7a9a5cba3c68fc59ac75
(cherry picked from commit 65558c7fcf4b45750891040ca27b6c2219262b32)