Query ports with admin client to get resource_request
The port.resource_request field is admin only. Nova depends on the
value of this field to do a proper scheduling and resource allocation
and deallocation for ports with resource request as well as to update
the port.binding:profile.allocation field with the resource providers
the requested resources are fulfilled from. However in some cases nova
does not use a neutron admin client / elevated context to read the
port. In this case neutron returns None for the port.resource_request
field and nova thinks that the port has no resource request.

This patch fixes all three places where previous testing showed that
context elevation was missing.

Change-Id: Icb35e20179572fb713a397b4605312cf3294b41b
Closes-Bug: #1945310