db: Remove models for removed services, features
We've built up a large amount of tables that are no longer used for
anything, given the removal of their users in past releases. We'd like
to remove those, but before we do that we need to drop the models. This
means there are no references to the tables come N+2, at which point we
can remove the tables themselves.

XenAPI virt driver
  Feature removed entirely in Victoria.

  - agent_builds
  - bw_usage_cache
  - console_pools
  - consoles

Cells v1
  Feature removed entirely in Train.

  - cells

Volume Snapshots
  Feature removed entirely in Liberty.

  - snapshots

EC2 API
  Feature removed entirely in Mitaka. Note that these tables are *not*
  used by the separate ec2-api project.

  - snapshot_id_mappings
  - volume_id_mappings

There are still some tables related to nova-network left here. Those are
unfortunately referenced from elsewhere, so we need to clean them up
separately.

Change-Id: I5e3d022fdf7328a1132f6e00998a3286b19be69a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>