db: Remove nova-network models
There are a number of nova-network tables which we can now drop.

nova-network
  Feature removed entirely in Ussuri.

  - dns_domains
  - fixed_ips
  - floating_ips
  - networks
  - provider_fw_rules
  - security_group_default_rules

Unfortunately we can't get rid of the security group-related entries due
to the unfortunate presence of the 'security_groups' attribute on the
'Instance' object and corresponding table, which in turn brings in a
load of other tables. We'll address that separately. For now, just drop
what we can easily drop.

Change-Id: I8858faa14119f4daa9630b0ff6dcf082d0ff8fba
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>