nova-manage: Ensure mountpoint is passed when updating attachment
This optional kwarg to the nova.volume.cinder.API.attachment_update
method ends up stashed in the connector passed to c-api and sets the
device associated with the attachment within Cinder. While this being
unset has no real world impact it should be kept the same as the
original attachment for completeness.

The Cinder fixture is extended to mimic the behaviour of
nova.volume.cinder.API.attachment_update prior to calling Cinder
allowing us to assert the value stashed in the connector and attachment
record.

Closes-Bug: #1945450
Change-Id: Ib2938a407598bf2dd466aae41700f350d2d34418