Vmware: Live-Migration of VMs
Tested with attached VMs for old cinder API and new (3.44+)
block-device-mapping.
Requires patches to cinder to "migrate" the volumes between vCenter by creating
an empty shadow-vm. Nova takes care of the rest
- deleting the outdated shadow-vms
- Attaching the volumes to the new shadow-vms

PlaceVM API is called to chose the host, but the initial placement doesn't take
the server groups into account, they are applied after the migration.

BigVMs are missing the probably needed special treatment (emptying of the host)

The code is written in a way, that if anything goes wrong during the migration,
such as
- the conductor
- one of the compute nodes
- one of the vCenter APIs
being offline/restarted, that simply a second migration can be started.

The migration may happen in the background of vCenter, and will be picked
up when finished. The necessary book-keeping will be updated after the
migration.

The migration data will only be kept in memory, and will not be recovered on
restart.

Since the VM is gone after a migration, it won't be possible to update
the vm-group membership and thus we do not call
`update_cluster_placement()` but only sync the server group instead.

We have to catch all exceptions in the post-migration steps, otherwise a
roll-back will be initiated which we cannot do properly as the VM has
already been migrated with the vSphere API.
Instead the VM will be set in error state as it will require
a manual inspection and intervention by an operator.

Apparently, the volume-id is not consistently stored as volume_id in
connection_info. Use the block_device.get_volume_id function to handle
the fallback.

Co-Authored-By: Jakob Karge <jakob.karge@sap.com>
Co-Authored-By: Marius Leustean <marius.leustean@sap.com>
Change-Id: If8c7265c53b64f00292e6689d1f6860ff29c671e
(cherry picked from commit ee0d89c7106a494948bf6d08f3285842b4e09e56)
Incorporated-Change-Id: I04c62f400a522bf9fe5828199c0dff80a1004f42
Incorporated-Change-Id: I75faecfdd48c9f40d243aecdd2b90b89e5158335
Incorporated-Change-Id: If5a8527578db8e4690595524e0785ee8b4de1d79
Incorporated-Change-Id: I192209d90603c43cb0c1bb4cb32e55c6638d9a3e